feat: Add Swiper
